+++ src/PermissionServiceProvider.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Spatie\Permission;
+namespace Yiyon\Permission;
 
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\Routing\Route;
 use Illuminate\Support\Collection;
 use Illuminate\Support\ServiceProvider;
 use Illuminate\View\Compilers\BladeCompiler;
-use Spatie\Permission\Contracts\Permission as PermissionContract;
-use Spatie\Permission\Contracts\Role as RoleContract;
+use Yiyon\Permission\Contracts\Permission as PermissionContract;
+use Yiyon\Permission\Contracts\Role as RoleContract;
 
 class PermissionServiceProvider extends ServiceProvider
 {