+++ tests/MiddlewareTest.php
@@ -1,16 +1,16 @@
 <?php
 
-namespace Spatie\Permission\Test;
+namespace Yiyon\Permission\Test;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use InvalidArgumentException;
-use Spatie\Permission\Contracts\Permission;
-use Spatie\Permission\Exceptions\UnauthorizedException;
-use Spatie\Permission\Middlewares\PermissionMiddleware;
-use Spatie\Permission\Middlewares\RoleMiddleware;
-use Spatie\Permission\Middlewares\RoleOrPermissionMiddleware;
+use Yiyon\Permission\Contracts\Permission;
+use Yiyon\Permission\Exceptions\UnauthorizedException;
+use Yiyon\Permission\Middlewares\PermissionMiddleware;
+use Yiyon\Permission\Middlewares\RoleMiddleware;
+use Yiyon\Permission\Middlewares\RoleOrPermissionMiddleware;
 
 class MiddlewareTest extends TestCase
 {