+++ CHANGELOG.md
@@ -162,7 +162,7 @@ https://github.com/spatie/laravel-permission/blob/master/config/permission.php
 2. Cache Resets: If your `app` or `tests` are clearing the cache by specifying the cache key, **it is better to use the built-in forgetCachedPermissions() method** so that it properly handles tagged cache entries. Here is the recommended change:
 ```diff
 - app()['cache']->forget('spatie.permission.cache');
-+ $this->app->make(\Spatie\Permission\PermissionRegistrar::class)->forgetCachedPermissions();
++ $this->app->make(\Yiyon\Permission\PermissionRegistrar::class)->forgetCachedPermissions();
 ```
 
 3. Also this is a good time to point out that now with v2.25.0 and v2.26.0 most permission-cache-reset scenarios may no longer be needed in your app, so it's worth reviewing those cases, as you may gain some app speed improvement by removing unnecessary cache resets.